chore(refactoring): code refactoring [Finishes #165409586]
