+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "start": "node build/server.js --inspect",
-    "start-dev": "nodemon --exec ./node_modules/.bin/babel-node ./src/server.js --inspect",
-    "build": "babel src --out-dir build",
+    "start-dev": "nodemon --exec ./node_modules/.bin/babel-node ./server/server.js --inspect",
+    "build": "babel server --out-dir build",
     "heroku-postbuild": "npm run build",
     "test": "nyc --reporter=text --reporter=lcov mocha --recursive --require @babel/polyfill --require @babel/register --exit",
     "coverage": "nyc --reporter=lcov --reporter=text mocha --recursive --require @babel/register --exit",
